This one work on several browsers
Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/ : false, ! what if I want to switch between both mode with the same code ? I can't act on the boolean? lack in flexibility and more power!

Failed to set the 'timeout' property on 'XMLHttpRequest': Timeouts cannot be set for synchronous requests made from a document.
:: same backgroud reason ! reduced options !

------------- the big missed of whatwg ?